[react] styling page

+++ frontend/src/components/form.js
@@ -1,7 +1,11 @@
 import { Component } from 'react'
 
 import React from 'react';
-import axios from 'axios'
+import axios from 'axios';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+
+import { faCheckCircle, faTrash, faXmark, faXmarkCircle } from '@fortawesome/free-solid-svg-icons'
+
 
 
 export class FormAdd extends Component {
@@ -93,16 +97,19 @@ export class FormAdd extends Component {
             })
     }
     deleteTodo = (id) => {
-        console.log(id)
-        axios.delete("http://localhost:8080/todos/" + id)
-            .then((response) => {
-                this.fetchTodos()
-                // console.log(response.data);
-                // console.log(this.state.todos)
-                // this.setState({ todos: response.data })
-                // console.log(this.state.todos)
+        if (window.confirm("Are you sure you want to delete this todo ?")) {
+            console.log(id)
+            axios.delete("http://localhost:8080/todos/" + id)
+                .then((response) => {
+                    this.fetchTodos()
+                    // console.log(response.data);
+                    // console.log(this.state.todos)
+                    // this.setState({ todos: response.data })
+                    // console.log(this.state.todos)
 
-            })
+                })
+        }
+
 
     }
     render() {
@@ -112,19 +119,20 @@ export class FormAdd extends Component {
             <div className='form-add'>
                 <section className=" get-in-touch">
                     <h1 className="title">Todo Application</h1>
-                    <form className="contact-form row" onSubmit={this.submitHandler}>
-                        <div className="form-field col-lg-6">
+
+                    <form className="contact-form row " onSubmit={this.submitHandler}>
+                        <div className="form-field col-lg-4">
                             <input id="title" className="input-text js-input" type="text" required value={title} name="title" onChange={this.changeHandler} />
                             <label className="label" for="title">Title</label>
                         </div>
-
-                        <div className="form-field col-lg-6 ">
+                        <br />
+                        <div className="form-field col-lg-4 ">
                             <input id="description" className="input-text js-input" type="text" value={description} name="description" onChange={this.changeHandler} required />
                             <label className="label" for="description">Description</label>
                         </div>
-                        <div className="form-field col-lg-6 ">
+                        <div className="form-field col-lg-4 ">
                             <select id="priority" className="input-text js-input" value={priority} name="priority"
-                                onChange={this.changeHandler} >
+                                onChange={this.changeHandler} required >
                                 <option value="" selected disabled hidden></option>
                                 <option value="HIGH">High</option>
                                 <option value="MEDIUM">Medium</option>
@@ -135,30 +143,45 @@ export class FormAdd extends Component {
                         </div>
 
 
-                        <div className="form-field col-lg-12">
+                        <div className="">
                             <button className="submit-btn" type="submit">Add</button>
                         </div>
                     </form>
                 </section>
 
                 <div className='list-todos get-in-touch table-responsive-md'>
-                    <div className="form-field col-lg-12" hidden={
+
+                    {/* <div className="form-field col-lg-12" hidden={
                         this.state.showUnfinished
                             ? true
                             : false
                     }>
                         <button className="submit-btn" type="submit" onClick={this.getUnfinishedTodos} >Get unfinished todos</button>
+                    </div> */}
+                    <div class="  pull-right float-right">
+                        <button class="btn btn-small btn-success filter" data-toggle="portfilter" data-target="all">
+                            All
+                        </button>
+                        <button class="btn btn-small btn-success filter" data-toggle="portfilter" data-target="art">
+                            Art
+                        </button>
+                        <button class="btn btn-small btn-success filter" data-toggle="portfilter" data-target="media">
+                            Media
+                        </button>
+
                     </div>
-                    <div className="form-field col-lg-12" hidden={
+
+                    {/* <div className="form-field col-lg-12" hidden={
                         this.state.showUnfinished
                             ? false
                             : true
-                    }>
-                        <button className="submit-btn" type="submit" onClick={this.fetchTodos} >Get all todos</button>
-                    </div>
-                    <div className="form-field col-lg-12" >
+                    }> */}
+                    {/* <button className="submit-btn" type="submit" onClick={this.fetchTodos} >Get all todos</button>
+                </div> */}
+                    {/* <div className="form-field col-lg-12" >
                         <button className="submit-btn" type="submit" onClick={this.getFinishedTodos} >Get finished todos</button>
-                    </div>
+                    </div> */}
+
 
                     <table className="table" style={{ width: 800 }}>
                         <thead className="black white-text">
@@ -192,19 +215,32 @@ export class FormAdd extends Component {
                                         } /></td>
                                     <td className={item.status ? 'checked-todo' : "none"} >{item.title}</td>
                                     <td className={item.status ? 'checked-todo' : "none"} >{item.description}</td>
-                                    <td className={item.status ? 'checked-todo' : "none"} >{item.priority}</td>
+                                    {item.priority == 'HIGH' &&
+                                        <td className={item.status ? 'checked-todo' : "none"} style={{ color: 'red' }}
+                                        >{item.priority}</td>}
+                                    {item.priority == 'MEDIUM' &&
+                                        <td className={item.status ? 'checked-todo' : "none"} style={{ color: 'orange' }}
+                                        >{item.priority}</td>}
+                                    {item.priority == 'LOW' &&
+                                        <td className={item.status ? 'checked-todo' : "none"} style={{ color: 'yellow' }}
+                                        >{item.priority}</td>}
+
                                     {item.status == 0 &&
                                         <td >
-                                            In progress
+                                            <FontAwesomeIcon icon={faXmarkCircle} color="red" />
+
+                                            {/* Unfinished */}
                                         </td>
                                     }
                                     {item.status == 1 &&
                                         <td className={item.status ? 'checked-todo' : "none"}>
-                                            Done
+                                            <FontAwesomeIcon icon={faCheckCircle} color="green" />
+                                            {/* Finished */}
                                         </td>
                                     }
                                     <td>
-                                        <button className='btn btn-primary' onClick={(e) => this.deleteTodo(item._id, e)}>Delete</button>
+
+                                        <button className='btn btn-danger' onClick={(e) => this.deleteTodo(item._id, e)} > <FontAwesomeIcon icon={faTrash} color="white" /> Delete</button>
                                     </td>
 
                                 </tr>
